+++ src/instructlab/model/evaluate.py
@@ -27,6 +27,7 @@ class Benchmark(str, enum.Enum):
     MMLU_BRANCH = "mmlu_branch"
     MT_BENCH = "mt_bench"
     MT_BENCH_BRANCH = "mt_bench_branch"
+    DK_BENCH = "dk_bench"
 
 
 def evaluate_model(
@@ -53,6 +54,10 @@ def evaluate_model(
     tls_client_passwd,  # pylint: disable=unused-argument
     enable_serving_output,
     skip_server: bool,
+    input_questions,
+    output_file_formats,
+    system_prompt,
+    temperature,
 ):
     """Evaluates a trained model"""
 
@@ -110,8 +115,25 @@ def evaluate_model(
         return evaluator.run(api_base)
 
     try:
-        # get appropriate evaluator class from Eval lib
+        # ensure judge_model and output_dir are not None and defaults are
+        # set correctly depending on benchmark being run.
+        if benchmark == Benchmark.MT_BENCH:
+            if judge_model is None:
+                judge_model = ctx.obj.config.evaluate.mt_bench.judge_model
+            if output_dir is None:
+                output_dir = ctx.obj.config.evaluate.mt_bench.output_dir
+        if benchmark == Benchmark.MT_BENCH_BRANCH:
+            if judge_model is None:
+                judge_model = ctx.obj.config.evaluate.mt_bench_branch.judge_model
+            if output_dir is None:
+                output_dir = ctx.obj.config.evaluate.mt_bench_branch.output_dir
+        elif benchmark == Benchmark.DK_BENCH:
+            if judge_model is None:
+                judge_model = ctx.obj.config.evaluate.dk_bench.judge_model
+            if output_dir is None:
+                output_dir = ctx.obj.config.evaluate.dk_bench.output_dir
 
+        # get appropriate evaluator class from Eval lib
         validate_options(
             model,
             base_model,
@@ -125,9 +147,49 @@ def evaluate_model(
             few_shots,
             batch_size,
             tasks_dir,
+            input_questions,
         )
 
-        if benchmark == Benchmark.MT_BENCH:
+        if benchmark == Benchmark.DK_BENCH:
+            # First Party
+            from instructlab.model.dk_bench_utils import (
+                print_results,
+                run_dk_bench,
+                validate_output_file_formats,
+                write_results,
+            )
+
+            # turn output_file_formats into a list to be passed into write_results
+            file_formats = output_file_formats.split(",")
+            validate_output_file_formats(file_formats)
+
+            # user did not pass in a system prompt
+            # setting default based on model
+            if system_prompt is None:
+                model_path = pathlib.Path(model)
+                system_prompt = get_sysprompt(get_model_arch(model_path))
+
+            result, model_name = run_dk_bench(
+                ctx,
+                model,
+                max_workers,
+                gpus,
+                backend,
+                enable_serving_output,
+                input_questions,
+                system_prompt,
+                temperature,
+                judge_model,
+            )
+
+            # default for output_dir is set by Click in src/instructlab/cli/model/evaluate.py
+            # it is a string not a pathlib.Path
+            files = write_results(result, file_formats, output_dir, model_name)
+            print_results(result, files, model_name)
+
+            logger.info("ᕦ(òᴗóˇ)ᕤ Model Evaluation with DK-Bench completed! ᕦ(òᴗóˇ)ᕤ")
+
+        elif benchmark == Benchmark.MT_BENCH:
             # Third Party
             from instructlab.eval.mt_bench import MTBenchEvaluator
 
@@ -417,6 +479,7 @@ def validate_options(
     few_shots,
     batch_size,
     tasks_dir,
+    input_questions,
 ):
     """takes in arguments from the CLI and uses 'benchmark' to validate other arguments
     if all needed configuration is present, raises an exception for the missing values
@@ -480,6 +543,40 @@ def validate_options(
         validate_model(model, allow_gguf=False)
         if benchmark == Benchmark.MMLU_BRANCH:
             validate_model(base_model, "--base-model", allow_gguf=False)
+
+    if benchmark == Benchmark.DK_BENCH:
+        required_args = [input_questions]
+        if None in required_args:
+            required_arg_names = ["input-questions"]
+            error_message = f"Benchmark {benchmark} requires the following args to be set: {required_arg_names}"
+            logger.error(f"\033[91m{error_message}\033[0m")
+            raise ValueError(error_message)
+
+        validate_output_dir(output_dir)
+
+
+def validate_output_dir(output_dir: str):
+    """
+    Validates that an output dir exists. If it does not it creates one.
+    If the input is a file, then a ValueError is thrown.
+
+    Args:
+        output_dir (str): Output directory for results
+    Returns:
+        None
+    """
+    output_dir_path = pathlib.Path(output_dir).resolve()
+    if not output_dir_path.exists():
+        logger.debug(
+            "Output directory %s does not exist. Creating %s .....",
+            output_dir_path,
+            output_dir_path,
+        )
+        os.makedirs(output_dir_path, exist_ok=True)
+    elif not output_dir_path.is_dir():
+        raise ValueError(
+            f"Output directory {output_dir_path} is a file not a directory",
+        )
 
 
 def validate_model(model: str, model_arg: str = "--model", allow_gguf: bool = True):